+++ tests/integration/conftest.py
@@ -1,10 +1,5 @@
-from unittest.mock import create_autospec
-
 import pytest
 from pyspark.sql import SparkSession
-
-from databricks.sdk import WorkspaceClient
-from databricks.sdk.service import iam
 
 
 @pytest.fixture(scope="session")
@@ -14,10 +9,3 @@ def mock_spark() -> SparkSession:
     :return: returns the spark session
     """
     return SparkSession.builder.appName("Remorph Reconcile Test").remote("sc://localhost").getOrCreate()
-
-
-@pytest.fixture()
-def mock_workspace_client():
-    client = create_autospec(WorkspaceClient)
-    client.current_user.me = lambda: iam.User(user_name="remorph", groups=[iam.ComplexValue(display="admins")])
-    yield client
