+++ tests/unit/conftest.py
@@ -22,14 +22,7 @@ from sqlglot import transpile
 
 from databricks.labs.remorph.config import SQLGLOT_DIALECTS, TranspileConfig
 from databricks.labs.remorph.reconcile.recon_config import (
-    ColumnMapping,
-    Filters,
-    JdbcReaderOptions,
     Schema,
-    Table,
-    ColumnThresholds,
-    Transformation,
-    TableThresholds,
 )
 from databricks.labs.remorph.transpiler.sqlglot.generator.databricks import Databricks
 from databricks.labs.remorph.transpiler.sqlglot.parsers.snowflake import Snowflake
@@ -212,64 +205,6 @@ def get_functional_test_files_from_directory(
     """Get all functional tests in the input_dir."""
     suite = parse_sql_files(input_dir, source, target, is_expected_exception)
     return suite
-
-
-@pytest.fixture
-def table_conf_mock():
-    def _mock_table_conf(**kwargs):
-        return Table(
-            source_name="supplier",
-            target_name="supplier",
-            jdbc_reader_options=kwargs.get('jdbc_reader_options', None),
-            join_columns=kwargs.get('join_columns', None),
-            select_columns=kwargs.get('select_columns', None),
-            drop_columns=kwargs.get('drop_columns', None),
-            column_mapping=kwargs.get('column_mapping', None),
-            transformations=kwargs.get('transformations', None),
-            column_thresholds=kwargs.get('thresholds', None),
-            filters=kwargs.get('filters', None),
-        )
-
-    return _mock_table_conf
-
-
-@pytest.fixture
-def table_conf_with_opts(column_mapping):
-    return Table(
-        source_name="supplier",
-        target_name="target_supplier",
-        jdbc_reader_options=JdbcReaderOptions(
-            number_partitions=100, partition_column="s_nationkey", lower_bound="0", upper_bound="100"
-        ),
-        join_columns=["s_suppkey", "s_nationkey"],
-        select_columns=["s_suppkey", "s_name", "s_address", "s_phone", "s_acctbal", "s_nationkey"],
-        drop_columns=["s_comment"],
-        column_mapping=column_mapping,
-        transformations=[
-            Transformation(column_name="s_address", source="trim(s_address)", target="trim(s_address_t)"),
-            Transformation(column_name="s_phone", source="trim(s_phone)", target="trim(s_phone_t)"),
-            Transformation(column_name="s_name", source="trim(s_name)", target="trim(s_name)"),
-        ],
-        column_thresholds=[
-            ColumnThresholds(column_name="s_acctbal", lower_bound="0", upper_bound="100", type="int"),
-        ],
-        filters=Filters(source="s_name='t' and s_address='a'", target="s_name='t' and s_address_t='a'"),
-        table_thresholds=[
-            TableThresholds(lower_bound="0", upper_bound="100", model="mismatch"),
-        ],
-    )
-
-
-@pytest.fixture
-def column_mapping():
-    return [
-        ColumnMapping(source_name="s_suppkey", target_name="s_suppkey_t"),
-        ColumnMapping(source_name="s_address", target_name="s_address_t"),
-        ColumnMapping(source_name="s_nationkey", target_name="s_nationkey_t"),
-        ColumnMapping(source_name="s_phone", target_name="s_phone_t"),
-        ColumnMapping(source_name="s_acctbal", target_name="s_acctbal_t"),
-        ColumnMapping(source_name="s_comment", target_name="s_comment_t"),
-    ]
 
 
 @pytest.fixture
